fix: run date adder on merge to main instead of PR
The problem is when a user submits a PR, we don't have access to run actions on their branch.(Testing was ran off own branches. Oops)

So I moved it into the merge to fix the conflict. Now the add-dates will run on merge of a PR to main since actions have permissions there.